feat(common/asset): move adminUI dev server to common
The adminUI bundle is not using a dependencyInjection configuration or extension.
