+++ EMS/admin-ui-bundle/config/services.xml
@@ -3,20 +3,12 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://symfony.com/schema/dic/services
         http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <parameters>
-        <parameter key="env(EMSCH_ENV)"/>
-    </parameters>
-
     <services>
         <defaults public="false" />
-        <service id="emsadminui.helper.dev_server" class="EMS\AdminUIBundle\Helper\Asset\DevServer">
-            <argument type="service" id="http_client"/>
-            <argument>%emsui.dev_server_url%</argument>
-        </service>
         <service id="emsadminui.helper.asset_version_strategy" class="EMS\AdminUIBundle\Helper\Asset\AssetVersionStrategy">
             <argument type="service" id="file_locator" />
             <argument type="service" id="request_stack" />
-            <argument type="service" id="emsadminui.helper.dev_server" />
+            <argument type="service" id="ems.asset.dev_server" />
             <tag name="twig.runtime" />
         </service>
     </services>
