--- EMS/admin-ui-bundle/src/DependencyInjection/Configuration.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EMS\AdminUIBundle\DependencyInjection;
-
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Builder\TreeBuilder;
-use Symfony\Component\Config\Definition\ConfigurationInterface;
-
-final class Configuration implements ConfigurationInterface
-{
-    #[\Override]
-    public function getConfigTreeBuilder(): TreeBuilder
-    {
-        $treeBuilder = new TreeBuilder('ems_admin_ui');
-        /* @var $rootNode ArrayNodeDefinition */
-        $rootNode = $treeBuilder->getRootNode();
-
-        $rootNode
-            ->children()
-                ->scalarNode('dev_server_url')->defaultNull()->end()
-            ->end()
-        ;
-
-        return $treeBuilder;
-    }
-}
