+++ EMS/common-bundle/config/services.xml
@@ -13,6 +13,11 @@
             <tag name="monolog.logger" channel="elastica" />
         </service>
 
+        <service id="ems.asset.dev_server" class="EMS\CommonBundle\Common\Asset\DevServer">
+            <argument type="service" id="http_client"/>
+            <argument>%ems_common.dev_server_url%</argument>
+        </service>
+
         <service id="ems.core_bridge.api" class="EMS\CommonBundle\Common\Bridge\Core\CoreApiBridge">
             <argument type="service" id="ems_common.core_api" />
         </service>
